+++ src/styles/index.js
@@ -1,0 +1,3 @@
+export { default as ThemeProvider } from './ThemeProvider';
+export { createPalette } from './palette';
+export { createTheme } from './theme';