+++ src/styles/palette.js
@@ -1,51 +1,56 @@
+import color from './color';
+
+const variant = {
+  primary: '#00A9EC',
+  info: '#365299',
+  success: '#00BD76',
+  warning: '#FFBD55',
+  danger: '#EF472F',
+};
+
 export const light = {
   text: {
-    primary: 'rgba(0, 0, 0, 0.87)',
-    secondary: 'rgba(0, 0, 0, 0.54)',
-  },
-  action: {
-    active: 'rgba(0, 0, 0, 0.54)',
-    disabled: 'rgba(0, 0, 0, 0.26)',
+    default: color.black,
+    muted: color.grayDarker,
   },
   background: {
-    default: '#ccc',
-    status: '#cc0',
+    default: color.white,
+    disabled: color.gray,
   },
 };
 
 export const dark = {
   text: {
-    primary: 'rgba(255, 255, 255, 0.87)',
-    secondary: 'rgba(255, 255, 255, 0.54)',
-  },
-  action: {
-    active: 'rgba(255, 255, 255, 0.54)',
-    disabled: 'rgba(255, 255, 255, 0.26)',
+    default: color.white,
+    secondary: color.grayLighter,
   },
   background: {
-    default: '#000',
-    status: '#cc0',
+    default: color.black,
+    disabled: color.grayDark,
   },
 };
 
 export const shades = { light, dark };
 
 export default function createPalette({
-  primary = 'rgba(75, 0, 130, 1)',
-  accent = 'rgba(255, 192, 203, 1)',
-  error = 'rgba(255, 0, 0, 1)',
-  type = 'light',
+  primary = variant.primary,
+  accent = variant.info,
+  error = variant.danger,
+  type = 'light', // TODO
+  name = 'nordnet',
 } = {}) {
   return {
+    name,
     type,
+    color,
+    variant,
     text: shades[type].text,
-    action: shades[type].action,
+    // action: shades[type].action,
     background: shades[type].background,
     shades,
     primary,
     accent,
     error,
-    // grey,
     // // functions
     // getContrastText,
   };