+++ src/test-utils/index.js
@@ -1,0 +1,9 @@
+import createMount from './createMount';
+import createShallow from './createShallow';
+import createRenderToString from './createRenderToString';
+
+export {
+  createMount,
+  createShallow,
+  createRenderToString,
+};